pyproject.toml
[tool.poetry]
name = "GameAI"
version = "0.1.0"
description = "Serpent Projects"
authors = ["Nicholas Shindler"]

[tool.poetry.dependencies]
python = "3.6"
cython = "0.26.1"
tensorflow = "^2.3"
numpy = "*"
keras = "^2.3"
opencv-python = "*"
tesserocr = "^2.5.1"


[tool.poetry.dev-dependencies]
wheel = "*"
setuptools = "*"
serpentai = "*"
pytest = "*"
pytest-cov = "*"
pytest-benchmark = "*"
redis = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
